focal/ubuntu-touch: add hooks to pre-install click'ed core apps
2 hooks are added: one that pulls from OpenStore, and one that pulls
from GitLab CI, which is intended to be temporary.

The Python code that pulls from OpenStore is copied from build-tools, so
that it happens as part of building rootfs itself instead of another CI
job. The code is slightly modified to remove the usage of `tqdm` library
which isn't in the rootfs.

Fixes: https://gitlab.com/ubports/development/core/rootfs-builder-debos/-/issues/12