focal: Don't replace existing /mnt directory with a symlink
/mnt might need to point to Android-specific files, but is also supposed to
be a directory as per default expectations of upstream Ubuntu.
Therefore it breaks Snap support in comical ways, due to AppArmor not being able
to mount a symlink into the target environment for the Snap app at hand.

A change ([1]) to lxc-android-config takes care of having /mnt still point to
the desired Android files.

1: https://gitlab.com/ubports/development/core/hybris-support/lxc-android-config/-/merge_requests/110