Pine64: Add SSH setup to the Pine overlay
This commit adds a ssh-generate-hostkeys in the Pine64 overlay and sets up the SSH
systemd service to call the script.
It will check whether the host keys are present, and generate them if not.
This is taken from the Halium lxc-android-config implementation, as that is not included
in the Pine64 RootFS.

Ultimately, this should probably be included in a package common to Halium and native
devices, but this commit would remove the need for Pine users to reinstall SSH.

Signed-off-by: Sumwun Ducainest <16982449-Sumwun-Ducainest@users.noreply.gitlab.com>